requirements.txt
# DO Architecture - Python Dependencies

# Core dependencies
python-dotenv>=1.0.0        # Environment variable management
requests>=2.31.0            # HTTP requests for API calls

# Apify integration
apify-client>=1.7.0         # Apify API client
fuzzywuzzy>=0.18.0          # Fuzzy string matching for validation
python-Levenshtein>=0.21.0  # Speed up fuzzy matching
openai>=1.0.0               # AI text generation

# Google APIs
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# Data processing
pandas>=2.1.0               # Data manipulation
openpyxl>=3.1.0            # Excel file handling

# Web scraping (optional, add if needed)
# beautifulsoup4>=4.12.0
# selenium>=4.15.0

# Rate limiting and retries
tenacity>=8.2.0            # Retry logic with exponential backoff

# Validation
email-validator>=2.1.0     # Email format validation

# Logging (optional enhancements)
# colorlog>=6.7.0          # Colored console output
